Break out branch and ref checks into module
