+++ lib/cog_cmd/cfn/template/show.rb
@@ -2,6 +2,8 @@ require 'cfn/command'
 
 module CogCmd::Cfn::Template
   class Show < Cfn::Command
+    include Cfn::RefOptions
+
     NAME_FORMAT = /\A[\w-]*\z/
 
     def run_command
@@ -29,18 +31,6 @@ module CogCmd::Cfn::Template
       end
     end
 
-    def require_ref_exists!
-      unless git_client.ref_exists?(ref)
-        if branch = ref[:branch]
-          raise(Cog::Error, "Branch #{branch} does not exist. Create a branch, push it to your repository's origin, and try again.")
-        elsif sha = ref[:tag]
-          raise(Cog::Error, "Tag #{tag} does not exist. Create a tag, push it to your repository's origin, and try again.")
-        elsif sha = ref[:sha]
-          raise(Cog::Error, "Git commit SHA #{sha} does not exist. Check that the SHA you are referencing has been pushed to your repository's origin and try again.")
-        end
-      end
-    end
-
     def require_template_exists!
       unless git_client.template_exists?(name, ref)
         if branch = ref[:branch]
@@ -57,30 +47,6 @@ module CogCmd::Cfn::Template
 
     def name
       request.args[0]
-    end
-
-    def ref
-      if branch
-        { branch: branch }
-      elsif tag
-        { tag: tag }
-      elsif sha
-        { sha: sha }
-      else
-        { branch: 'master' }
-      end
-    end
-
-    def branch
-      request.options['branch']
-    end
-
-    def tag
-      request.options['tag']
-    end
-
-    def sha
-      request.options['sha']
     end
   end
 end
